* updated GL internal format detection in ImageDataRenderTarget::initRenderTarget()  * changed EEPGenerator render target texture format to GL_RGBA16
git-svn-id: https://camplinux.in.tum.de/svn/campvis/trunk@260 bb408c1c-ae56-11e1-83d9-df6b3e0c105e

+++ src/tumvis/modules/vis/eepgenerator.cpp
@@ -93,7 +93,7 @@ namespace TUMVis {
                 glEnable(GL_CULL_FACE);
 
                 // create entry points texture
-                ImageDataRenderTarget* entrypoints = new ImageDataRenderTarget(tgt::svec3(_renderTargetSize.getValue(), 1), GL_RGBA16F);
+                ImageDataRenderTarget* entrypoints = new ImageDataRenderTarget(tgt::svec3(_renderTargetSize.getValue(), 1), GL_RGBA16);
                 entrypoints->activate();
 
                 glDepthFunc(GL_LESS);
@@ -105,7 +105,7 @@ namespace TUMVis {
                 entrypoints->deactivate();
 
                 // create exit points texture
-                ImageDataRenderTarget* exitpoints = new ImageDataRenderTarget(tgt::svec3(_renderTargetSize.getValue(), 1), GL_RGBA16F);
+                ImageDataRenderTarget* exitpoints = new ImageDataRenderTarget(tgt::svec3(_renderTargetSize.getValue(), 1), GL_RGBA16);
                 exitpoints->activate();
 
                 glDepthFunc(GL_GREATER);
